+++ gaios/Receive/ReceiveViewController.swift
@@ -656,7 +656,7 @@ extension ReceiveViewController: UITableViewDelegate, UITableViewDataSource {
                 case .backup:
                     cell.configure(alertCard,
                                    onLeft: {[weak self] in
-                        if let vc = AccountNavigator.recover() {
+                        if let vc = AccountNavigator.backupIntro(.quiz) {
                             self?.navigationController?.pushViewController(vc, animated: true)
                         }
                     },