+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added create account functionality in settings tab
 
 ### Changed
-- Wallet Scurity Tab: changed illustrations
+- Wallet Security Tab: changed illustrations
+- Improved backup flow instructions
 
 ## [5.1.1] - 2025-10-15